requirements.txt
py-cord==2.4.0
aiohttp==3.8.3
apscheduler==3.10.1
dateutils==0.6.12
numpy
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.28.2